+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- avoid js snippet execution in native context
 
 ## 12.14.0 - 2021/1/24