+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+
+## 12.14.1 - 2021/1/27
+
 - fix bug when dom-capture was taken with incorrect scroll location during `extractText` operation
 - avoid js snippet execution in native context
 - updated to @applitools/driver@1.0.4 (from 1.0.2)